created reusable contest filter component
