+++ pages/index.tsx
@@ -3,6 +3,7 @@ import Layout from '../components/Layout'
 import { getContestList } from '../services/reventService'
 import React, { useState, useEffect } from 'react'
 import SearchInput from '../components/SearchInput'
+import ContestFilter from '../components/ContestFilter'
 
 export default function Home(props) {
   const { contestList } = props
@@ -14,17 +15,6 @@ export default function Home(props) {
   const filteredContestList = contestList.filter(
     (contest) => statusFilter === 'All' || contest.status === statusFilter
   )
-
-  const handleStatusFilter = (status: string) => {
-    setStatusFilter(status)
-  }
-
-  useEffect(() => {
-    // When hiding the filters, reset them
-    if (open == false) {
-      setStatusFilter('All')
-    }
-  }, [open])
 
   return (
     <Layout>
@@ -39,50 +29,12 @@ export default function Home(props) {
             </button>
 
             <SearchInput />
-            {open && (
-              <div>
-                <button
-                  className={
-                    statusFilter === 'All'
-                      ? 'text-orange-500 bg-orange-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                      : 'text-gray-700 bg-gray-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                  }
-                  onClick={() => handleStatusFilter('All')}
-                >
-                  All
-                </button>
-                <button
-                  className={
-                    statusFilter === 'Open'
-                      ? 'text-orange-500 bg-orange-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                      : 'text-gray-700 bg-gray-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                  }
-                  onClick={() => handleStatusFilter('Open')}
-                >
-                  Open
-                </button>
-                <button
-                  className={
-                    statusFilter === 'Voting'
-                      ? 'text-orange-500 bg-orange-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                      : 'text-gray-700 bg-gray-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                  }
-                  onClick={() => handleStatusFilter('Voting')}
-                >
-                  Voting
-                </button>
-                <button
-                  className={
-                    statusFilter === 'Closed'
-                      ? 'text-orange-500 bg-orange-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                      : 'text-gray-700 bg-gray-700 text-white px-3 py-2 rounded-2xl font-medium cursor-pointer mr-2 mt-2'
-                  }
-                  onClick={() => handleStatusFilter('Closed')}
-                >
-                  Closed
-                </button>
-              </div>
-            )}
+            <ContestFilter
+              contestList={contestList}
+              open={open}
+              statusFilter={statusFilter}
+              setStatusFilter={setStatusFilter}
+            />
           </div>
           <main className='min-h-screen py-8 px-20 flex-1 flex flex-col'>
             <div className='grid grid-cols-4 gap-4'>
